server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
google-api-python-client==2.149.0
pdfplumber==0.11.4
Pillow==10.4.0
python-dotenv==1.0.1